requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
croniter==1.4.1
deprecation==2.1.0
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-WTF==1.1.1
frozenlist==1.4.0
gotrue==1.0.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.21.1
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.9
packaging==23.1
postgrest==0.10.8
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==3.2.0
realtime==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tqdm==4.66.1
twilio==8.7.0
typing_extensions==4.7.1
urllib3==2.0.4
websockets==10.4
Werkzeug==2.3.7
WTForms==3.0.1
yarl==1.9.2
